fixing test, implementing messages for recipe searching
