+++ src/component/Route.js
@@ -4,7 +4,6 @@ import { NavigationContext } from "../context/NavigationContext";
 function Route({ children, path }) {
   const { currentPath } = useContext(NavigationContext);
   const re = new RegExp(path);
-
   if (currentPath.match(re)) {
     return children;
   }
