+++ src/context/NavigationContext.js
@@ -1,4 +1,4 @@
-import { useState, createContext, useEffect } from 'react';
+import { useState, createContext, useEffect } from "react";
 
 const NavigationContext = createContext();
 function NavigationProvider({ children }) {
@@ -8,15 +8,15 @@ function NavigationProvider({ children }) {
     const handler = () => {
       setCurrentPath(window.location.pathname);
     };
-    window.addEventListener('popstate', handler);
+    window.addEventListener("popstate", handler);
 
     return () => {
-      window.removeEventListener('popstate', handler);
+      window.removeEventListener("popstate", handler);
     };
   }, []);
 
   const navigate = (to) => {
-    window.history.pushState({}, '', to);
+    window.history.pushState({}, "", to);
     setCurrentPath(to);
   };
